requirements.txt
Django==2.2.3
django-geojson==2.12.0
django-leaflet==0.24.0
jsonfield==2.0.2
Pillow==6.1.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
